LPD-47337 Rename ContentFDSPropsTransformer to ContentsFDSPropsTransformer
