+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/contents_section.jsp
@@ -21,7 +21,7 @@ ContentsSectionDisplayContext contentsSectionDisplayContext = (ContentsSectionDi
 		formName="fm"
 		id="<%= CMSSiteInitializerFDSNames.CONTENTS_SECTION %>"
 		itemsPerPage="<%= 10 %>"
-		propsTransformer="{ContentFDSPropsTransformer} from site-cms-site-initializer"
+		propsTransformer="{ContentsFDSPropsTransformer} from site-cms-site-initializer"
 		selectedItemsKey="id"
 		selectionType="multiple"
 		style="fluid"
