+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/index.ts
@@ -3,6 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as ContentFDSPropsTransformer} from './FDSPropsTransformer/ContentFDSPropsTransformer';
+export {default as ContentsFDSPropsTransformer} from './FDSPropsTransformer/ContentsFDSPropsTransformer';
 export {default as FilesFDSPropsTransformer} from './FDSPropsTransformer/FilesFDSPropsTransformer';
 export {default as StructureBuilder} from './structure_builder/components/StructureBuilder';
